+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
 ### Latest/Nightly (this branch)
+- Features
+    - User messages now link out to there respective frame and message. #2
 - Bug Fixes
     - Fixed selecting non-packet frame not removing frame detail panel. closes #1
+- Internal
+    - set source-demo-tool version to 0.4.1.
+        - Affects UserMessagesToolViewModel, addressed in message index feature.
 
 ### v0.3.0
 - Internal